rename Cell class
